Fix 32-bit

+++ test/struct.jl
@@ -131,7 +131,7 @@ end
 struct J
     id::Union{Int, Nothing}
     name::Union{String, Nothing}
-    rate::Union{Int, Float64}
+    rate::Union{Int64, Float64}
 end
 
 struct K
